Added kvm_init and KVM put/get regs for ARM
Also added a few debug statements, which should be removed
later on.

The code proceeds all the way to KVM_RUN, where it crashes.
It crashes in the PrepareSwitchToGuest function, which
must be fixed!

Next items are fixing the above, and then eventually IRQ
injection and looking at memory things.

Also, fixed the configure script to have the right include
statement in kvm_cflags for ARM in the Makefiles.